chore(formatters): move formatter code into util file - the method `autoAddEditorFormatterToColumnsWithEditor` with previous code could not be used by other framework (like Aurelia-Slickgrid) because it was not in a common section folder, so this PR simply moves it to a common area which all framework (not only slick-vanilla-bundle) can use and avoid duplicate code
