+++ packages/common/src/formatters/formatterUtilities.ts
@@ -2,7 +2,35 @@ import { FieldType } from '../enums/fieldType.enum';
 import { Column, ExcelExportOption, Formatter, GridOption, SlickGrid, TextExportOption } from '../interfaces/index';
 import { mapMomentDateFormatWithFieldType } from '../services/utilities';
 import * as moment_ from 'moment-mini';
+import { multipleFormatter } from './multipleFormatter';
 const moment = (moment_ as any)['default'] || moment_; // patch to fix rollup "moment has no default export" issue, document here https://github.com/rollup/rollup/issues/670
+
+/**
+ * Automatically add a Custom Formatter on all column definitions that have an Editor.
+ * Instead of manually adding a Custom Formatter on every column definitions that are editables, let's ask the system to do it in an easier automated way.
+ * It will loop through all column definitions and add an Custom Editor Formatter when necessary,
+ * also note that if there's already a Formatter on the column definition it will automatically use the Formatters.multiple and add the custom formatter into the `params: formatters: {}}`
+ */
+export function autoAddEditorFormatterToColumnsWithEditor(columnDefinitions: Column[], customEditableFormatter: Formatter) {
+  if (Array.isArray(columnDefinitions)) {
+    for (const columnDef of columnDefinitions) {
+      if (columnDef.editor) {
+        if (columnDef.formatter && columnDef.formatter !== multipleFormatter && columnDef.formatter !== customEditableFormatter) {
+          const prevFormatter = columnDef.formatter;
+          columnDef.formatter = multipleFormatter;
+          columnDef.params = { ...columnDef.params, formatters: [prevFormatter, customEditableFormatter] };
+        } else if (columnDef.formatter && columnDef.formatter === multipleFormatter && columnDef.params) {
+          // before adding the formatter, make sure it's not yet in the params.formatters list, we wouldn't want to add it multiple times
+          if (columnDef.params.formatters.findIndex((formatter: Formatter) => formatter === customEditableFormatter) === -1) {
+            columnDef.params.formatters = [...columnDef.params.formatters, customEditableFormatter];
+          }
+        } else {
+          columnDef.formatter = customEditableFormatter;
+        }
+      }
+    }
+  }
+}
 
 /**
  * Find the option value from the following (in order of execution)
